Add shredded and iteration files
Update from the V1_simultated_tactile_stimuli file. Now, trial number can be modified. The shredded file has a lot of visualizations to modify a single variable and see the effects on all the outputs. The iterations file allows the visualization of error dependency versus a specific variable, for example, nFolds or nRepeats.